net: migrate CreateMonsterFromStoreForGC() to CanonicalCookies::Create method
This patch converts the usage of CanonicalCookie's constructor in
CreateMonsterFromStoreForCG() to CanonicalCookie's Create() method.

We are doing this because Create() is more robust, as it does
validation of the parameters and return NULL as an error signal
if they fail to comply with the checks.

BUG=57061
TEST=net_unittests
R=mmenke@chromium.org

Review-Url: https://codereview.chromium.org/2030233002
Cr-Original-Commit-Position: refs/heads/master@{#397771}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aa0b9f052e9e4a0ac0ec38a41ba44c8be1f1bb16