Add |pkp_bypassed| to SSLInfo.
This is the first step in exposing in devtools when pinning is bypassed due to a local trust anchor. To correctly expose this for all socket types, |pkp_bypassed| is also exposed in a CertificateVerifyResult, and set by each socket after calling TransportSecurityState::CheckPublicKeyPins.

This also updates HttpResponseInfo to serialize the new field correctly.

BUG=566144

Review-Url: https://codereview.chromium.org/2016143002
Cr-Original-Commit-Position: refs/heads/master@{#399242}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: df302c4d9de0bcd65378ecd44a4e3ec1199dea4c