net: migrate SQLitePersistentCookieStoreTest to CanonicalCookies::Create method
This patch converts the usage of CanonicalCookie's constructor in SQLitePersistentCookieStoreTest
to CanonicalCookie's Create() method.

We are doing this because Create() is more robust, as it does validation of the parameters and
return NULL as an error signal if they fail to comply with the checks.

BUG=57061
TEST=net_unittests
R=mmenke@chromium.org

Review-Url: https://codereview.chromium.org/2065643002
Cr-Original-Commit-Position: refs/heads/master@{#399586}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 594b1aef628787dba01ed765810dfe2e422e007a