net: migrate SQLitePersistentCookieStoreTest to CanonicalCookies::Create method - part 2
This patch converts the remaining the usage of CanonicalCookie's constructor in SQLitePersistentCookieStoreTest
to CanonicalCookie's Create() method.

We are doing this because Create() is more robust, as it does validation of the parameters and
return NULL as an error signal if they fail to comply with the checks.

BUG=57061
TEST=net_unittests
R=mmenke@chromium.org

Review-Url: https://codereview.chromium.org/2069753002
Cr-Original-Commit-Position: refs/heads/master@{#400024}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f84b59e65ea9c24c94dafeaf1e960b43f9dd2b93