Remove the test-only X509Certificate constructor
net::X509Certificate had a constructor that was intended
for test-only code that allowed minting a fake
certificate. However, this is a frequent source of test
bugs, because these certificates violate one of
X509Certificate's core guarantees - that there's always
an associated OSCertHandle for an X509Certificate.

This removes the ctor and all usages of it. Code either
imports a test certificate directly or mints them on the
fly, as appropriate.

BUG=526899
TBR=rdevlin.cronin@chromium.org

Review-Url: https://codereview.chromium.org/2084153003
Cr-Original-Commit-Position: refs/heads/master@{#401430}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f300d683274a040d68a367aeed7aecbdc3a611a8