Ensure native library isn't actually loaded, and no IO on net thread
LoadNativeLibrary asserts because IO is possible, so call dlopen()
directly with flag ensuring no IO actually happens. We need to
access libnetd_client.so, which is guaranteed to be loaded by the
point where it's needed as socket() is guaranteed to be called
which libnetd_client.so handles.

BUG=623555
R=xunjieli

Review-Url: https://codereview.chromium.org/2104533002
Cr-Original-Commit-Position: refs/heads/master@{#402817}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 25ea5bc3b93a92ced13ac18936a34053521c4bca