Synchronize events of CFNetworkExecuteProxyAutoConfigurationURL
CFNetworkExecuteProxyAutoConfigurationURL run loop sources are not
thread safe. Therefore, the following source events should be
synchronized:
1. Adding the source to the run loop.
2. Handling the result of proxy resolution.
3. Removing the source from the run loop.

This change does not prevent the parallel resolution of proxies for
multiple URLs but only prevents concurrent execution of the above
mentioned event, which are supposed to be lightweight operations.

BUG=166387

Review-Url: https://codereview.chromium.org/2094373002
Cr-Original-Commit-Position: refs/heads/master@{#403284}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 19a49ad54e88393e509771ad0659dd6004849140