Certificate Transparency: MerkleTreeLeaf, MerkleAuditProof improvements.
Per comments on another code review, I've documented MerkleTreeLeaf and
changed the Hash function name.

This change also adds the tree size to the MerkleAuditProof, as each
proof ties to a particular tree size (and it can't directly be
deduced from the number of nodes in the proof).

BUG=613495

Review-Url: https://codereview.chromium.org/2107423004
Cr-Original-Commit-Position: refs/heads/master@{#404356}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b38ec0a4aaf590b21c1344a5962bcfeb65f3e3d6