Remove URLRequest::Delegate::OnBeforeNetworkStart.
URLRequest::Delegate::OnBeforeNetworkStart is not used by
anything, despite having a lot of plumping.  It was added to
try experimenting with having the ResourceScheduler delay
requests only once they needed to go over the network, so
responses that could be served out of the cache.

BUG=479994

Review-Url: https://codereview.chromium.org/2123093002
Cr-Original-Commit-Position: refs/heads/master@{#404606}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0af4009f7ce7a5a00091070e85bfd36745bba02c