Add CHECK to make sure QuicHttpStream does not get destroyed during DoLoop
A fewer crashers on QuicHttpStream::DoReadRequestBody
indicate that QuicHttpStream might be destroyed in
the middle of DoLoop. The crashers are pretty rare. This CL
adds a bool so that if QuicHttpStream gets destroyed during
DoLoop, we will get the call stack which will tell us how
it is being destroyed.

BUG=627099

Review-Url: https://codereview.chromium.org/2140673006
Cr-Original-Commit-Position: refs/heads/master@{#405328}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1f48e6ebda2ca94af20d049aff0863244232ffc5