Add the ability to send Expect-Staple reports.
This adds a |ProcessExpectStaple| method to TransportSecurityState that
implements reporting for Expect-Staple, using contents of SSLInfo and
OCSPVerifyResult. No report is sent if a valid staple is provided, or if
the remote host is not on the Expect-Staple preload list. This method is
not currently called by any socket implementation.

BUG=598021

Review-Url: https://codereview.chromium.org/2144693004
Cr-Original-Commit-Position: refs/heads/master@{#406742}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c87b456d823a4bd61599c5ce4512848a6f2be7b6