Do not invoke QuicHttpStream::DoCallback when in DoLoop
Latest crash dumps suggest that callback is invoked
re-entrantly during QuicHttpStream::DoLoop(). This can be
problematic because the caller can delete the
QuicHttpStream when the callback is run. This CL changes
QuicHttpStream::OnClose() to not invoke callback when
already in DoLoop().

This CL also adds a unit test which will crash without the
change or the CHECKs.

BUG=629043

Review-Url: https://codereview.chromium.org/2163883004
Cr-Original-Commit-Position: refs/heads/master@{#407152}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8dff50bad755b439be74e98225647699f555ee96