Prevents connection migration from crashing on asynchronous writes
Problem: We observed in the wild that connection migration will
occasionally crash chrome shortly after the migration completes.

Solution: The problem is that connection migration creates a new packet
writer that should be initialized before use.  Since it was not
initialized, the chrome packet writer would calls OnWriteComplete on its
connection_ object, which was a nullptr.

We now call Initialize on the new packet writer.  Added a
corresponding unittest that performs an asynchronous write, verified
that it did exercise the bug, and is now fixed.  Also added a DCHECK
to prevent obvious nullptr dereferences.

R=rch@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2228683004
Cr-Original-Commit-Position: refs/heads/master@{#410796}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c330e85e9138e8c8ce6446494802190ca03fcb52