Misc SSLClientSocket cleanup inspired by NPN removal.
Remove GetNextProto(), because callers do not care if the negotiated protocol
was also on the server list or is the client fallback protocol (besides NPN has
been disabled for a while anyway).  They can just call GetNegotiatedProtocol()
instead.  This makes SSLSocketDataProvider::next_proto_status unused, so remove
that too.

Move GetNegotiatedProtocol() and WasNpnNegotiated() implementation from
SSLClientSocket to derived classes.  This must be done because they cannot rely
on GetNextProto() any longer.  This is cleaner though, because there is one
fewer layer.

Remove NextProtoStatus argument from NetLogHttpStreamProtoCallback, because now
that NPN is disabled, it does not carry useful information.

Rename SSLClientSocketImpl::npn_proto_ to negotiated_protocol_, because not only
NPN can be used to negotiate it.  In fact, NPN has been disabled for some time
now, so this rename is long overdue.

Change SSLSocketDataProvider::next_proto and
SSLClientSocketImpl::negotiated_protocol_ type from std::string to NextProto,
because logically that is the kind of information they hold.

Inline uses of now trivial SSLSocketDataProvider::SetNextProto().

Move RecordNegotiationExtension(), RecordChannelIDSupport(), and
IsChannelIDEnabled() from SSLClientSocket to SSLClientSocketImpl, because that
is the only derived class that calls them.  Also make them not static.

Move negotiation_extension_ from SSLClientSocket to SSLClientSocketImpl, because
that is the only derived class that uses it.  This member will be removed very
soon.

Change NextProtoFromString() to take base::StringPiece instead of std::string,
so that BoringSSL results do not have to be copied.

BUG=526713

Review-Url: https://codereview.chromium.org/2219133002
Cr-Original-Commit-Position: refs/heads/master@{#411327}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3cf2a5904ddc4228c15f9a540a556f93e216918c