Fixed a nullptr dereference in the connection migration histogram logic
Basically, we were clossing the connection, and then subsequently
dereferencing its pointer to get its connection_id for the Histogram.

Swapped the order of the lines so that we histogram the connection_id
prior to clossing the connection.

I did not add a unittest for the scenario because it's not clear how
to make this happen, even in the wild.  When a session migrates, it is
marked as going away.  Once it goes away, you can't make it active
again, so it can't migrate.  As a consequence, it's not clear how
we'll ever hit the "too many migrations" scenario.

I filed a bug to re-visit this issue in the future.  For now, it's
sufficient to just prevent a possible nullptr dereference.

R=rch@chromium.org
BUG=

Review-Url: https://codereview.chromium.org/2240943002
Cr-Original-Commit-Position: refs/heads/master@{#412001}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9bfcb872cdae84c05b23cc8006678866849f0a3b