Removes TestVector from ct_log_verifier_unittest
Its only purpose was to bundle together a hex string and the number of
bytes it represents. The number of bytes (length_bytes) is redundant
though, as it is easily calculated by strlen(str) / 2, and is not used
anyway. The one thing that appeared to use it was HexToBytes, but this
contained a bug that meant it both misinterpreted and ignored it.

This also optimizes GetEmptyTreeHash by only calculating it once, which
reduces total test duration by ~1 second.

BUG=631087

Review-Url: https://codereview.chromium.org/2275353002
Cr-Original-Commit-Position: refs/heads/master@{#414552}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 540ba6cbe242920544a1f43019e4f4a3c6f81bf8