Add a network quality cache observer
The observer is notified every time there is a change in the
cached network quality. In the later CLs, one such observer
would be created per-profile. On receiving the notification,
it would write the network id and the cached network quality
(both are included in the notification) to the prefs.

BUG=490870

Review-Url: https://codereview.chromium.org/2261813002
Cr-Original-Commit-Position: refs/heads/master@{#414815}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 24d7cace893a94082f913ddf0a1dbc6f725dda31