Race TCP connections to proxies with QUIC connections
After the default job finishes proxy resolution, it queries
ProxyDelegate on whether the resolved proxy supports an
alternative proxy server or not. If yes, a new job is
started. This job is provided the alternative proxy server,
and races with the main job.

If the alternative proxy server is found to be broken, the
ProxyDelegate is notified.

Changes to DataReductionProxyDelegate will be in a
subsequent CL.

BUG=343579

Review-Url: https://codereview.chromium.org/2260623002
Cr-Original-Commit-Position: refs/heads/master@{#414884}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c3308d79532cd022c3394d6f3b0773958307d2be