Non-functional simplification of NetworkChangeNotifierAutoDetect code
Turns out NetworkState was only needed for the current default network,
so simplify unnecessary pieces.  This is in preparation for adding
WiFi SSID to NetworkState, so I wanted to reduce the dependence on
NetworkState so the code wouldn't be required to fetch the Wifi
SSID as much.

BUG=647691
R=xunjieli

Review-Url: https://codereview.chromium.org/2361563003
Cr-Original-Commit-Position: refs/heads/master@{#420653}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e30b8a04dde1f68f8a03f915bf448e65bbd7a19