Add net::BrotliSourceStream
This CL adds a net::BrotliSourceStream to do brotli decoding.
The new class is based on net::BrotliFilter but implements the
new pull-based interface of net::FilterSourceStream.

This is a part of the effort to convert net::Filter into a pull-based
interface. See the linked bug for more details.

BUG=474859

Review-Url: https://codereview.chromium.org/2329973003
Cr-Original-Commit-Position: refs/heads/master@{#420664}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 084a9290cededdcaf31e6ce30c8f58875fb34198