Make SSLClientSessionCache a client of memory coordinator
When the memory coordinator is enabled, SSLClientSessionCache becomes a
client of the memory coordinator instead of installing
MemoryPressureListener.

In the implmentation of this CL, both MemoryPressureListener and
MemoryCoordinatorClient are installed regardless of the flag, but as
default MemoryCoordinator is not enabled and when MemoryCoordinator is
enabled by a flag, it is planed to suppress MemoryPressureMonitor
features.

Design Doc: https://docs.google.com/document/d/1a69mMr7jI7qK0OfKNlrZ350xiXizVMCCe8orGX7K8Uo/edit?ts=57d7968b#

BUG=639700

Review-Url: https://codereview.chromium.org/2384703002
Cr-Original-Commit-Position: refs/heads/master@{#422697}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e10e500bba646fdeb77b2ae6b9b4927d311968d9