Make SdchOwner a client of memory coordinator
When the memory coordinator is enabled, SdchOwner becomes a client of
the memory coordinator instead of installing MemoryPressureListener.

In the implmentation of this CL, both MemoryPressureListener and
MemoryCoordinatorClient are installed regardless of the flag, but as
default MemoryCoordinator is not enabled and when MemoryCoordinator is
enabled by a flag, it is planed to suppress MemoryPressureMonitor
features.

Design Doc: https://docs.google.com/document/d/1a69mMr7jI7qK0OfKNlrZ350xiXizVMCCe8orGX7K8Uo/edit?ts=57d7968b#

BUG=639700

Review-Url: https://codereview.chromium.org/2390603002
Cr-Original-Commit-Position: refs/heads/master@{#422698}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 651a10b49a97e21f805aca384d734f3879cd66db