Remove OnDataAvailable() notification when trailers are to be delivered
This CL used the change that is suggested by mpw@ in
https://bugs.chromium.org/p/chromium/issues/detail?id=652935#c1.

Specifically, this CL removes
QuicChromiumClientStream::Delegate::OnDataAvailable()
notification if there is no byte can be read and trailers are to be delivered.

This CL adjusted unit tests to test the correct ordering of events.

BUG=652935

Review-Url: https://codereview.chromium.org/2393233002
Cr-Original-Commit-Position: refs/heads/master@{#423357}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7640faa369a7769b97719938c70e4e4910f476c3